pyproject.toml
[tool.poetry]
name = "medical-records-api"
version = "0.1.0"
description = "Medical Records AI Analysis API with FastAPI"
authors = ["Mark Rahimi <markrahimi@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
motor = "^3.6.0"
pymongo = "^4.10.0"
pydantic-settings = "^2.6.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.12"
groq = "^0.11.0"
email-validator = "^2.2.0"
resend = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "PT"]
ignore = ["E501", "B008"]
